+++ src/dbus/settingsadaptor/settingsadaptor_registry_scrolling.cpp
@@ -733,6 +733,8 @@ void SettingsAdaptor::initializeRegistryScrolling()
         m_schemas[QStringLiteral("scrollingStickyWindowHandling")] = QStringLiteral("int");
         REGISTER_CONCRETE_BOOL("scrollingRespectMinimumSize", scrollingRespectMinimumSize,
                                setScrollingRespectMinimumSize)
+        REGISTER_CONCRETE_BOOL("scrollingCenterShortColumns", scrollingCenterShortColumns,
+                               setScrollingCenterShortColumns)
         REGISTER_CONCRETE_BOOL("scrollingRestoreStripsOnLogin", scrollingRestoreStripsOnLogin,
                                setScrollingRestoreStripsOnLogin)
         // scrollingRestoreFloatedWindowsOnLogin and scrollingKeepFloatingAbove are